+++ apps/server/src/modules/account/controller/dto/password-patter.spec.ts
