+++ apps/server/src/modules/account/controller/dto/password-pattern.ts
