+++ apps/server/src/modules/account/services/dto/account.dto.ts
@@ -1,42 +1,17 @@
 import { EntityId } from '@shared/domain';
+import { AccountSaveDto } from './account.save.dto';
 
-export class AccountDto {
-	id: EntityId;
+export class AccountDto extends AccountSaveDto {
+	readonly id: EntityId;
 
-	createdAt: Date;
+	readonly createdAt: Date;
 
-	updatedAt: Date;
-
-	username: string;
-
-	password?: string;
-
-	token?: string;
-
-	credentialHash?: string;
-
-	userId: EntityId;
-
-	systemId?: EntityId;
-
-	lasttriedFailedLogin?: Date;
-
-	expiresAt?: Date;
-
-	activated?: boolean;
+	readonly updatedAt: Date;
 
 	constructor(props: AccountDto) {
+		super(props);
 		this.id = props.id;
 		this.createdAt = props.createdAt;
 		this.updatedAt = props.updatedAt;
-		this.username = props.username;
-		this.password = props.password;
-		this.token = props.token;
-		this.credentialHash = props.credentialHash;
-		this.userId = props.userId;
-		this.systemId = props.systemId;
-		this.lasttriedFailedLogin = props.lasttriedFailedLogin;
-		this.expiresAt = props.expiresAt;
-		this.activated = props.activated;
 	}
 }
