--- apps/server/src/modules/authentication/mapper/account-entity-to-dto.mapper.ts
@@ -1,21 +1,0 @@
-import { Account } from '@shared/domain';
-import { AccountDto } from '../services/dto/account.dto';
-
-export class AccountEntityToDtoMapper {
-	static mapToDto(account: Account): AccountDto {
-		return new AccountDto({
-			id: account.id,
-			createdAt: account.createdAt,
-			updatedAt: account.updatedAt,
-			userId: account.userId,
-			username: account.username,
-			activated: account.activated,
-			credentialHash: account.credentialHash,
-			expiresAt: account.expiresAt,
-			lasttriedFailedLogin: account.lasttriedFailedLogin,
-			password: account.password,
-			systemId: account.systemId,
-			token: account.token,
-		});
-	}
-}
